requirements.txt
streamlit
opencv_python==4.5.1.48
np
Pillow
face_recognition